Move TestEmbeddings and TestModel to internals
