+++ js/packages/teams-ai/src/embeddings/index.ts
@@ -8,4 +8,3 @@
 
 export * from './EmbeddingsModel';
 export * from './OpenAIEmbeddings';
-export * from './TestEmbeddings';
