+++ js/packages/teams-ai/src/internals/testing/TestModel.ts
@@ -7,10 +7,11 @@
  */
 
 import { TurnContext } from 'botbuilder';
-import { Message, PromptFunctions, PromptTemplate } from '../prompts';
-import { PromptCompletionModel, PromptResponse, PromptResponseStatus } from './PromptCompletionModel';
-import { Tokenizer } from '../tokenizers';
-import { Memory } from '../MemoryFork';
+
+import { Memory } from '../../MemoryFork';
+import { PromptCompletionModel, PromptResponse, PromptResponseStatus } from '../../models/PromptCompletionModel';
+import { Message, PromptFunctions, PromptTemplate } from '../../prompts';
+import { Tokenizer } from '../../tokenizers';
 
 /**
  * A test model that can be used to test the prompt completion system.
@@ -18,9 +19,9 @@ import { Memory } from '../MemoryFork';
 export class TestModel implements PromptCompletionModel {
     /**
      *
-     * @param status Optional. Status of the prompt response. Defaults to `success`.
-     * @param response Optional. Response to the prompt. Defaults to `{ role: 'assistant', content: 'Hello World' }`.
-     * @param error Optional. Error to return. Defaults to `undefined`.
+     * @param {PromptResponseStatus} status Optional. Status of the prompt response. Defaults to `success`.
+     * @param {Message} response Optional. Response to the prompt. Defaults to `{ role: 'assistant', content: 'Hello World' }`.
+     * @param {Error} error Optional. Error to return. Defaults to `undefined`.
      */
     public constructor(
         status: PromptResponseStatus = 'success',
@@ -49,12 +50,12 @@ export class TestModel implements PromptCompletionModel {
 
     /**
      * Completes a prompt.
-     * @param context Current turn context.
-     * @param memory An interface for accessing state values.
-     * @param functions Functions to use when rendering the prompt.
-     * @param tokenizer Tokenizer to use when rendering the prompt.
-     * @param template Prompt template to complete.
-     * @returns A `PromptResponse` with the status and message.
+     * @param {TurnContext} context Current turn context.
+     * @param {Memory} memory An interface for accessing state values.
+     * @param {PromptFunctions} functions Functions to use when rendering the prompt.
+     * @param {Tokenizer} tokenizer Tokenizer to use when rendering the prompt.
+     * @param {PromptTemplate} template Prompt template to complete.
+     * @returns {Promise<PromptResponse<string>>} A `PromptResponse` with the status and message.
      */
     public async completePrompt(
         context: TurnContext,
